+++ datanode/entities/account_balance.go
@@ -29,21 +29,11 @@ type AccountBalance struct {
 }
 
 func (ab *AccountBalance) ToProto() *v2.AccountBalance {
-	marketID := ab.MarketID.String()
-	if marketID == noMarketStr {
-		marketID = ""
-	}
-
-	ownerID := ab.PartyID.String()
-	if ownerID == systemOwnerStr {
-		ownerID = ""
-	}
-
 	return &v2.AccountBalance{
-		Owner:    ownerID,
+		Owner:    ab.PartyID.String(),
 		Balance:  ab.Balance.String(),
 		Asset:    ab.AssetID.String(),
-		MarketId: marketID,
+		MarketId: ab.MarketID.String(),
 		Type:     ab.Account.Type,
 	}
 }